+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "with-typescript",
+  "name": "QBACC Home Project",
   "version": "1.0.0",
   "scripts": {
     "dev": "next",
