+++ pages/api/users/index.ts
@@ -1,5 +1,16 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import { sampleUserData } from "../../../utils/sample-data";
+
+export type User = {
+  id: number;
+  name: string;
+};
+
+export const sampleUserData: User[] = [
+  { id: 101, name: "Alice" },
+  { id: 102, name: "Bob" },
+  { id: 103, name: "Caroline" },
+  { id: 104, name: "Dave" },
+];
 
 const handler = (_req: NextApiRequest, res: NextApiResponse) => {
   try {
